dataviz-enhanced/requirements.txt
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.1.0
numpy>=1.26.0
openpyxl>=3.1.0
Pillow>=10.0.0
PyYAML>=6.0
beautifulsoup4>=4.12.0
scipy>=1.11.0
cairosvg>=2.7.0
